refactor: Convert task list into a component
refactor: Convert task list into a component

* Moved all task list related code from Home.js to TaskList.js
* Imported TaskList into Home.js
* Home.js now uses <TaskList> component call
