+++ front-end/src/pages/Home.js
@@ -8,8 +8,7 @@ import Card from "react-bootstrap/Card";
 import DatePicker from "react-datepicker";
 import "react-datepicker/dist/react-datepicker.css";
 import Calendar from "react-calendar";
-import ListGroup from "react-bootstrap/ListGroup";
-import InputGroup from "react-bootstrap/InputGroup";
+import TaskList from "../components/TaskList.js"
 
 const Home = () => {
   const [taskName, setNameInput] = useState(""); // New state for the name input
@@ -36,7 +35,7 @@ const Home = () => {
         .then((res) => {
           res.json();
         })
-        .then((data) => {});
+        .then((data) => { });
 
       console.log("Successfully added task!");
     } else {
@@ -72,146 +71,16 @@ const Home = () => {
       .then((data) => setTasks(data));
   }, []);
 
-  function CheckBox({ id, taskID, taskName, taskDate, taskDesc, taskStatus }) {
-    let labelID = id + "label";
-    let checkedID = id + "checked";
-
-    if (taskStatus.toLowerCase() === "complete") {
-      return (
-        <InputGroup>
-          <Form.Check
-            checked="checked"
-            onChange={handleChecked(
-              id,
-              labelID,
-              taskID,
-              taskName,
-              taskDate,
-              taskDesc,
-              taskStatus
-            )}
-            type="checkbox"
-            id={checkedID}
-          />
-          <label className={"checked"} id={labelID} htmlFor={id}>
-            {taskName}
-          </label>
-        </InputGroup>
-      );
-    }
-    return (
-      <InputGroup>
-        <Form.Check
-          onChange={handleChecked(
-            id,
-            labelID,
-            taskID,
-            taskName,
-            taskDate,
-            taskDesc,
-            taskStatus
-          )}
-          type="checkbox"
-          id={id}
-        />
-        <label className={"unchecked"} id={labelID} htmlFor={id}>
-          {taskName}
-        </label>
-      </InputGroup>
-    );
-  }
-
-  const handleChecked =
-    (checkID, labelID, taskID, taskName, taskDate, taskDesc, taskStatus) =>
-    (event) => {
-      let label = document.getElementById(labelID);
-
-      if (checkID.includes("checked")) {
-        return;
-      }
-
-      if (label.className === "checked") {
-        taskStatus = "Incomplete";
-        label.className = "unchecked";
-      } else if (label.className === "unchecked") {
-        taskStatus = "Complete";
-        label.className = "checked";
-      }
-      let fetchRequest =
-        "https://planned-out-backend-jdx6.onrender.com/updatetask/:" + taskID;
-
-      // TODO: Mark the task as incomplete in the database
-      fetch(fetchRequest, {
-        method: "PUT",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({
-          id: taskID,
-          name: taskName,
-          date: taskDate,
-          desc: taskDesc,
-          status: taskStatus,
-        }),
-      })
-        .then((res) => {
-          res.json();
-        })
-        .then((data) => {});
-
-      console.log("Successfully updated task!");
-    };
-
-  const [tasksList, setTaskList] = useState([]);
-
-  useEffect(() => {
-    fetch("https://planned-out-backend-jdx6.onrender.com/")
-      .then((res) => res.json())
-      .then((data) => setTaskList(data));
-  }, []);
-
   return (
     <div className="App">
       <Container>
         <Row>
           <Col>
             {/* Tasks Card */}
-            {/* Tasks Card */}
             <Card className="tasks-card">
               <Card.Title>Today's Tasks</Card.Title>
               {/* List of Tasks */}
-              <ListGroup className="task-list">
-                <ListGroup.Item className="task">
-                  <InputGroup>
-                    <Form.Check
-                      onClick={handleChecked("testtask", "testtasklabel")}
-                      type="checkbox"
-                      id="testtask"
-                    />
-                    <label
-                      className={"unchecked"}
-                      id="testtasklabel"
-                      htmlFor="testtask"
-                    >
-                      This is a test task
-                    </label>
-                  </InputGroup>
-                </ListGroup.Item>
-                {/* For loop for each task in the tasks list */}
-                {tasksList.map((task) => (
-                  <ListGroup.Item className="task">
-                    <CheckBox
-                      className="task"
-                      id={"task" + tasksList.indexOf(task)}
-                      taskID={task._id}
-                      taskName={task.taskName}
-                      taskDate={task.taskDate}
-                      taskDesc={task.taskDesc}
-                      taskStatus={task.taskStatus}
-                    ></CheckBox>
-                  </ListGroup.Item>
-                ))}
-              </ListGroup>
+              <TaskList />
               {/* Spacer */}
               <div className="d-flex flex-column"></div>
               {/* Add a task button */}
@@ -268,9 +137,8 @@ const Home = () => {
                   tileContent={({ date }) => {
                     if (tasks) {
                       // Convert the date to a string in "MM-DD-YYYY" format
-                      const formattedDate = `${
-                        date.getMonth() + 1
-                      }-${date.getDate()}-${date.getFullYear()}`;
+                      const formattedDate = `${date.getMonth() + 1
+                        }-${date.getDate()}-${date.getFullYear()}`;
 
                       // Check if there is a task for the selected date
                       const hasTask = tasks.some(
